Added to post and socket logic
Posting to the /testing route now saves the zip file out and returns an id.
The client can call the socket event 'test' and pass it the ID to start the extraction of the zip and run the tests.
A simple index page was added to see the results.
The data is returned by the socket and displayed on the index page.