Updated testing route logic
Started creating a real temporary folder which is destroyed when the process is closed.
Added an 'init' function export to the module glob. Init is run once when the express app is setting up.
Got spawn to work spinning up mocha and streaming it successfully to the socket.
Included the process close signal which signals the end of the tests and streams how many tests failed.
Piped mocha's standard error output to the servers error output for debugging purposes.