Changed the logic behind how the tests work
Tests are no longer extracted when ran. When they are sent over as zips, they are extracted and kept, no longer destroyed.
Test projects now send over an ID which identifies which project they belong to.
status emission is now type safe and linked to the front-end's type.
Logic has been cut down to save on sanity.
meta.json files are created and read when tests are ran to return old fetch data.