Updated the logic for sending and testing modules.
The socket now emits a module object instead of a raw result.
A new Meta class was added which handles updating, creating and saving the meta.json.
When the server extracts a zip, the meta.json will be created and the last updated date is pushed.
Cleaned up the source file a bit.
Imported definitions from the client side into defs.ts