Feature/mojo testing (#13)
* Started creating the layout for the testing framework

Added SocketIO and JsZip.
Added a route for testing.
Setup Socket.IO as a module.

* Added file upload capabilities and test data.

Added SocketIO.
Created a "room" interface for mixing IO channels.
Created a post and socket listener for the testing route.
Created logic to store the posted zip files to the temp folder for extraction and evaluation.

* Added to post and socket logic

Posting to the /testing route now saves the zip file out and returns an id.
The client can call the socket event 'test' and pass it the ID to start the extraction of the zip and run the tests.
A simple index page was added to see the results.
The data is returned by the socket and displayed on the index page.

* Added screenshots and tests temporary folders to git.

Folders need to exist in order to work for now, will likely have a server startup task to create folders if not exists.

* Updated testing route logic

Started creating a real temporary folder which is destroyed when the process is closed.
Added an 'init' function export to the module glob. Init is run once when the express app is setting up.
Got spawn to work spinning up mocha and streaming it successfully to the socket.
Included the process close signal which signals the end of the tests and streams how many tests failed.
Piped mocha's standard error output to the servers error output for debugging purposes.

* Updated temp/test cleanup logic

Added a clean directory method call to the initialization of the route.
Updated the 'how.md' containing the brief explaination of the system.
Isolated 'tmp' into TempDir for reuse.

* removed the test for '.zip' in the file name

I'm moved the file to a blob, so it comes back as blob.

* Fixed issues with temporary directory routing.

had a seperate path for zip and extracted.
add event off calls to prevent leakage.

* File path resolving for files.

Added access control allow origin headers.
Added a download get request that posts the test zip.

* Started changing logic to handle project naming.

Removed the init function that cleared the test folder on restart.
Started writing notes to convert the extraction and creation process to include the project ID handed over.
Started writing notes for creating a meta.json file that stores the previous test metadata for the project.

* Changed the logic behind how the tests work

Tests are no longer extracted when ran. When they are sent over as zips, they are extracted and kept, no longer destroyed.
Test projects now send over an ID which identifies which project they belong to.
status emission is now type safe and linked to the front-end's type.
Logic has been cut down to save on sanity.
meta.json files are created and read when tests are ran to return old fetch data.

* Updated the logic for sending and testing modules.

The socket now emits a module object instead of a raw result.
A new Meta class was added which handles updating, creating and saving the meta.json.
When the server extracts a zip, the meta.json will be created and the last updated date is pushed.
Cleaned up the source file a bit.
Imported definitions from the client side into defs.ts